refactor: moving abacate pay billing and customer to a service
