+++ src/main/java/com/abacatepay/model/IAbacatePay.java
@@ -1,5 +1,9 @@
 package com.abacatepay.model;
 
+import com.abacatepay.services.IAbacatePayBilling;
+import com.abacatepay.services.IAbacatePayCustomer;
+
 public interface IAbacatePay {
     IAbacatePayBilling billing();
+    IAbacatePayCustomer customer();
 }
