+++ src/main/java/com/abacatepay/services/IAbacatePayCustomer.java
@@ -1,0 +1,10 @@
+package com.abacatepay.services;
+
+import com.abacatepay.model.customer.CreateCustomerResponse;
+import com.abacatepay.model.customer.CustomerMetadata;
+import com.abacatepay.model.customer.ListCustomerResponse;
+
+public interface IAbacatePayCustomer {
+    CreateCustomerResponse create(CustomerMetadata data);
+    ListCustomerResponse list();
+}
