+++ src/test/java/com/abacatepay/model/customer/CustomerMetadataTest.java
@@ -7,11 +7,12 @@ class CustomerMetadataTest {
 
     @Test
     void testGettersAndSetters() {
-        CustomerMetadata customerMetadata = new CustomerMetadata();
-        customerMetadata.setName("John Doe");
-        customerMetadata.setCellphone("123456789");
-        customerMetadata.setEmail("john.doe@example.com");
-        customerMetadata.setTaxId("12345678901");
+        CustomerMetadata customerMetadata = CustomerMetadata.builder()
+                .name("John Doe")
+                .cellphone("123456789")
+                .email("john.doe@example.com")
+                .taxId("12345678901")
+                .build();
 
         assertEquals("John Doe", customerMetadata.getName(), "Name field must be set correctly");
         assertEquals("123456789", customerMetadata.getCellphone(), "Cellphone field must be set correctly");
